2025.7.31 Push PNG images to the screen
An Image URL entity provides a mechanism to pass the url of a PNG image (that can be to your local home assistant www folder or to an online image) that will be displayed on the screen for 60 seconds. This can be leveraged in home assistant automations. Images will be resized appropriately and will be removed after they have been shown.