2025.12.21 Weather and Substitutions
Merry Christmas! This release brings significant changes in the form of substitutions making it easier than ever to customise and deploy this configuration. Simply update the relevant substitutions to tailor the configuration - only entities which report a valid state will be shown on the device. The screensaver can also now show weather from an external source.